+++ facebook_auth_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+### 2.6.0
+- Removed LoginBehavior and FacebookAuthErrorCode class.
+- Added enum LoginBehavior.
+- Improved tests.
 ### 2.5.0
 - Added isWebSdkInitialized;